requirements.txt
Django==4.0.4
django-mptt==0.13.4
django-colorfield==0.6.3

